+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reworked custom email sending.
 - Made to/from addresses in emails pretty.
 - Made the site domain in emails use the site framework domain, so make sure it's set properly.
+- Updated default logging settings.
 
 ### Deprecated